fix(CourseCard): move CourseCard to Core folder
