+++ client/src/components/Core/CourseCard/CourseCard.module.scss
