+++ client/src/views/Authorized/Participants/taalhuis/Participants/Detail/LearningNeeds/Details/Tests/Detail/ParticipantsLearningNeedsReferencesTestUpdateView.tsx
@@ -17,7 +17,7 @@ import Modal from 'components/Core/Modal/Modal'
 import ModalView from 'components/Core/Modal/ModalView'
 import SectionTitle from 'components/Core/Text/SectionTitle'
 import Paragraph from 'components/Core/Typography/Paragraph'
-import CourseCard from 'components/Domain/Taalhuis/CourseCard/CourseCard'
+import CourseCard from 'components/Core/CourseCard/CourseCard'
 import { TaalhuisParticipantLearningNeedReferenceFields } from 'components/Domain/Taalhuis/TaalhuisLearningNeedsReferenceCreateFields'
 import { TaalhuisParticipantLearningNeedReferenceTestFields } from 'components/Domain/Taalhuis/TaalhuisLearningNeedsReferenceTestFields'
 import { LearningOutcomeOfferFieldsetModel } from 'components/fieldsets/participants/learningNeeds/fieldsets/LearningOutcomeOfferFieldset'
